add assertXpathQuery* methods, move files to keep organisations, remove unused namespace
