+++ library/Zend/Test/PHPUnit/Controller/AbstractControllerTestCase.php
@@ -12,11 +12,9 @@ namespace Zend\Test\PHPUnit\Controller;
 
 use PHPUnit_Framework_TestCase;
 use PHPUnit_Framework_ExpectationFailedException;
-use Zend\Dom;
 use Zend\EventManager\StaticEventManager;
 use Zend\Http\Request as HttpRequest;
 use Zend\Mvc\Application;
-use Zend\ModuleManager\ModuleEvent;
 use Zend\Mvc\MvcEvent;
 use Zend\Mvc\View\SendResponseListener;
 use Zend\Stdlib\Exception\LogicException;
@@ -632,234 +630,5 @@ abstract class AbstractControllerTestCase extends PHPUnit_Framework_TestCase
             ));
         }
         $this->assertNotEquals($route, $match);
-    }
-
-    /**
-     * Execute a dom query
-     * @param string $path
-     * @return array
-     */
-    protected function query($path)
-    {
-        $response = $this->getResponse();
-        $dom = new Dom\Query($response->getContent());
-        $result = $dom->execute($path);
-        return $result;
-    }
-
-    /**
-     * Count the dom query executed
-     * @param string $path
-     * @return integer
-     */
-    protected function queryCount($path)
-    {
-        return count($this->query($path));
-    }
-
-    /**
-     * Assert against DOM selection
-     *
-     * @param  string $path CSS selector path
-     * @return void
-     */
-    public function assertQuery($path)
-    {
-        $match = $this->queryCount($path);
-        if(!$match > 0) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s EXISTS', $path
-            ));
-        }
-        $this->assertEquals(true, $match > 0);
-    }
-
-    /**
-     * Assert against DOM selection
-     *
-     * @param  string $path CSS selector path
-     * @return void
-     */
-    public function assertNotQuery($path)
-    {
-        $match = $this->queryCount($path);
-        if($match != 0) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s DOES NOT EXIST', $path
-            ));
-        }
-        $this->assertEquals(0, $match);
-    }
-
-    /**
-     * Assert against DOM selection; should contain exact number of nodes
-     *
-     * @param  string $path CSS selector path
-     * @param  string $count Number of nodes that should match
-     * @return void
-     */
-    public function assertQueryCount($path, $count)
-    {
-        $match = $this->queryCount($path);
-        if($match != $count) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s OCCURS EXACTLY %d times, actually occurs %d times',
-                $path, $count, $match
-            ));
-        }
-        $this->assertEquals($match, $count);
-    }
-
-    /**
-     * Assert against DOM selection; should NOT contain exact number of nodes
-     *
-     * @param  string $path CSS selector path
-     * @param  string $count Number of nodes that should NOT match
-     * @return void
-     */
-    public function assertNotQueryCount($path, $count)
-    {
-        $match = $this->queryCount($path);
-        if($match == $count) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s DOES NOT OCCUR EXACTLY %d times',
-                $path, $count
-            ));
-        }
-        $this->assertNotEquals($match, $count);
-    }
-
-    /**
-     * Assert against DOM selection; should contain at least this number of nodes
-     *
-     * @param  string $path CSS selector path
-     * @param  string $count Minimum number of nodes that should match
-     * @return void
-     */
-    public function assertQueryCountMin($path, $count)
-    {
-        $match = $this->queryCount($path);
-        if($match < $count) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s OCCURS AT LEAST %d times, actually occurs %d times',
-                $path, $count, $match
-            ));
-        }
-        $this->assertEquals(true, $match >= $count);
-    }
-
-    /**
-     * Assert against DOM selection; should contain no more than this number of nodes
-     *
-     * @param  string $path CSS selector path
-     * @param  string $count Maximum number of nodes that should match
-     * @return void
-     */
-    public function assertQueryCountMax($path, $count)
-    {
-        $match = $this->queryCount($path);
-        if($match > $count) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s OCCURS AT MOST %d times, actually occurs %d times',
-                $path, $count, $match
-            ));
-        }
-        $this->assertEquals(true, $match <= $count);
-    }
-
-    /**
-     * Assert against DOM selection; node should contain content
-     *
-     * @param  string $path CSS selector path
-     * @param  string $match content that should be contained in matched nodes
-     * @return void
-     */
-    public function assertQueryContentContains($path, $match)
-    {
-        $result = $this->query($path);
-        if($result->count() == 0) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s EXISTS', $path
-            ));
-        }
-        if($result->current()->nodeValue != $match) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node denoted by %s CONTAINS content "%s", actual content is "%s"',
-                $path, $match, $result->current()->nodeValue
-            ));
-        }
-        $this->assertEquals($result->current()->nodeValue, $match);
-    }
-
-    /**
-     * Assert against DOM selection; node should NOT contain content
-     *
-     * @param  string $path CSS selector path
-     * @param  string $match content that should NOT be contained in matched nodes
-     * @return void
-     */
-    public function assertNotQueryContentContains($path, $match)
-    {
-        $result = $this->query($path);
-        if($result->count() == 0) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s EXISTS', $path
-            ));
-        }
-        if($result->current()->nodeValue == $match) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s DOES NOT CONTAIN content "%s"',
-                $path, $match
-            ));
-        }
-        $this->assertNotEquals($result->current()->nodeValue, $match);
-    }
-
-    /**
-     * Assert against DOM selection; node should match content
-     *
-     * @param  string $path CSS selector path
-     * @param  string $pattern Pattern that should be contained in matched nodes
-     * @return void
-     */
-    public function assertQueryContentRegex($path, $pattern)
-    {
-        $result = $this->query($path);
-        if($result->count() == 0) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s EXISTS', $path
-            ));
-        }
-        if(!preg_match($pattern, $result->current()->nodeValue)) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node denoted by %s CONTAINS content MATCHING "%s", actual content is "%s"',
-                $path, $pattern, $result->current()->nodeValue
-            ));
-        }
-        $this->assertEquals(true, (boolean)preg_match($pattern, $result->current()->nodeValue));
-    }
-
-    /**
-     * Assert against DOM selection; node should NOT match content
-     *
-     * @param  string $path CSS selector path
-     * @param  string $pattern pattern that should NOT be contained in matched nodes
-     * @return void
-     */
-    public function assertNotQueryContentRegex($path, $pattern)
-    {
-        $result = $this->query($path);
-        if($result->count() == 0) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s EXISTS', $path
-            ));
-        }
-        if(preg_match($pattern, $result->current()->nodeValue)) {
-            throw new PHPUnit_Framework_ExpectationFailedException(sprintf(
-                'Failed asserting node DENOTED BY %s DOES NOT CONTAIN content MATCHING "%s"',
-                $path, $pattern
-            ));
-        }
-        $this->assertEquals(false, (boolean)preg_match($pattern, $result->current()->nodeValue));
     }
 }
